ubuntu-requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
benzinga==1.21
blinker==1.4
boto3==1.28.55
botocore==1.31.55
Brlapi==0.8.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.4
command-not-found==0.3
contourpy==1.1.1
cryptography==41.0.4
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.1
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
duplicity==0.8.21
effdet==0.4.1
emoji==2.8.0
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.103.1
fasteners==0.14.1
filelock==3.12.4
filetype==1.2.0
FLAML==2.1.1
flatbuffers==23.5.26
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.2
google-api-core==2.13.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
greenlet==3.0.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==4.6.4
iopath==0.1.10
ipykernel==6.25.2
ipython==8.16.1
jaraco.context==4.3.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.4.0
jupyter_core==5.4.0
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.0.302
langdetect==1.0.9
langsmith==0.0.41
language-selector==0.1
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
loguru==0.7.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openai==0.28.1
opencv-python==4.8.1.78
packaging==23.1
pandas==2.1.1
paramiko==2.9.3
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
peewee==3.17.0
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==3.11.0
playwright==1.39.0
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyautogen==0.1.11
pycairo==1.20.1
pycocotools==2.0.7
pycparser==2.21
pycups==2.0.1
pydantic==1.10.9
pydantic_core==2.10.1
pyee==11.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyngrok==7.0.0
pyparsing==2.4.7
pypdfium2==4.21.0
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.4.0
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.0
scipy==1.11.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.3
starlette==0.27.0
structlog==23.2.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.5.1
timm==0.9.7
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.0
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
union==0.1.10
unstructured==0.10.24
unstructured-inference==0.7.7
unstructured.pytesseract==0.3.12
uritemplate==4.1.1
urllib3==1.25.10
usb-creator==0.3.7
uvicorn==0.23.2
wadllib==1.3.6
wcwidth==0.2.8
webencodings==0.5.1
wolframalpha==5.0.0
xdg==5
xkit==0.0.0
xmltodict==0.13.0
yarl==1.9.2
yfinance==0.2.31
zipp==1.0.0
